pull request

+++ Proga/fastPow/Matrix.cpp
@@ -6,13 +6,13 @@
 		//init(n, m);
 		n = n1;
 		m = m1;
-		data = new int*[n];
-		newData = new int*[n];
+		data = new unsigned long long*[n];
+		newData = new unsigned long long*[n];
 
 		for (int i = 0; i < n; i++)
 		{
-			data[i] = new int[m];
-			newData[i] = new int[m];
+			data[i] = new unsigned long long[m];
+			newData[i] = new unsigned long long[m];
 			for (int j = 0; j < m; j++)
 			{
 				data[i][j] = 0;
@@ -21,37 +21,37 @@
 		}
 	}
 
-	Matrix::Matrix(int** matrix, int n1, int m1)
+	Matrix::Matrix(unsigned long long** matrix, int n1, int m1)
 	{
 		n = n1;
 		m = m1;
-		data = new int*[n];
-		newData = new int*[n];
+		data = new unsigned long long*[n];
+		newData = new unsigned long long*[n];
 
-		for (int i = 0; i < n; i++)
+		for (unsigned long long i = 0; i < n; i++)
 		{
-			data[i] = new int[m];
-			newData[i] = new int[m];
-			for (int j = 0; j < m; j++)
+			data[i] = new unsigned long long[m];
+			newData[i] = new unsigned long long[m];
+			for (unsigned long long j = 0; j < m; j++)
 			{
 				data[i][j] = 0;
 				newData[i][j] = 0;
 			}
 		}
-		for (int i = 0; i < n; i++)
+		for (unsigned long long i = 0; i < n; i++)
 		{
-			for (int j = 0; j < m; j++)
+			for (unsigned long long j = 0; j < m; j++)
 				data[i][j] = matrix[i][j];
 		}
 	}
 
 	Matrix::Matrix()
 	{
-		for (int i = 0; i < n; i++)
+		for (unsigned long long i = 0; i < n; i++)
 		{
-			data[i] = new int[m];
+			data[i] = new unsigned long long[m];
 			if (!newData[i])
-				newData[i] = new int[m];
+				newData[i] = new unsigned long long[m];
 			for (int j = 0; j < m; j++)
 			{
 				data[i][j] = 0;
@@ -62,16 +62,16 @@
 	Matrix* Matrix::mul(Matrix &other)
 	{
 		assert(this->m == other.n);
-		newData = new int*[this->m];
-		for (int i = 0; i < n; i++)
-			newData[i] = new int[other.n];
+		newData = new unsigned long long*[this->m];
+		for (unsigned long long i = 0; i < n; i++)
+			newData[i] = new unsigned long long[other.n];
 
-		for (int i = 0; i < this->n; i++)
+		for (unsigned long long i = 0; i < this->n; i++)
 		{
-			for (int j = 0; j < other.m; j++)
+			for (unsigned long long j = 0; j < other.m; j++)
 			{
 				newData[i][j] = 0;
-				for (int k = 0; k < this->m; k++)
+				for (unsigned long long k = 0; k < this->m; k++)
 				{
 					newData[i][j] += data[i][k] * other.data[k][j];
 				}
@@ -84,7 +84,7 @@
 
 	Matrix::~Matrix()
 	{
-		for (int i = 0; i < n; i++)
+		for (unsigned long long i = 0; i < n; i++)
 		{
 			if (data[i])
 				delete[] data[i];
